pyproject.toml
[project]
name = "my-pinboard"
version = "0.1.0"
description = "Personal Pinboard clone"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "jinja2",
    "python-multipart",
    "requests",
    "beautifulsoup4",
    "trafilatura",
    "pydantic",
    "playwright>=1.56.0",
    "google-generativeai>=0.8.5",
    "google-genai>=1.52.0",
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
